Setup of systems

+++ Context/ProcessMachinePartContext.cs
@@ -2,10 +2,10 @@ namespace ToolGame.Context;
 
 public class ProcessMachinePartContext
 {
-    public long Delta;
+    public double Delta;
     public Machine3D Machine;
 
-    public ProcessMachinePartContext(long delta, Machine3D machine)
+    public ProcessMachinePartContext(double delta, Machine3D machine)
     {
         Delta = delta;
         Machine = machine;
